server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
faster-whisper>=1.0.0
python-multipart==0.0.6
